+++ packages/terradart_google/lib/dns.dart
@@ -23,3 +23,5 @@ export 'src/dns/google_dns_managed_zone.dart'
         DnsManagedZonePrivateVisibilityNetwork;
 export 'src/dns/google_dns_managed_zone_iam_member.dart'
     show GoogleDnsManagedZoneIamMember;
+export 'src/dns/google_dns_policy.dart' show GoogleDnsPolicy;
+export 'src/dns/google_dns_record_set.dart' show GoogleDnsRecordSet;