+++ packages/terradart_google/lib/eventarc.dart
@@ -6,6 +6,8 @@ library;
 
 export 'src/eventarc/google_eventarc_channel.dart'
     show GoogleEventarcChannel;
+export 'src/eventarc/google_eventarc_google_channel_config.dart'
+    show GoogleEventarcGoogleChannelConfig;
 export 'src/eventarc/google_eventarc_enrollment.dart'
     show GoogleEventarcEnrollment;
 export 'src/eventarc/google_eventarc_google_api_source.dart'